requirements.txt
PyGithub
pandas
markdown
docutils
Pygments
python-dotenv
black
loguru
joblib